+++ packages/graphql/src/schema-model/attribute/model-adapters/AttributeAdapter.ts
@@ -108,7 +108,8 @@ export class AttributeAdapter {
             this.typeHelper.isUserScalar() ||
             this.typeHelper.isEnum() ||
             this.typeHelper.isTemporal() ||
-            this.typeHelper.isSpatial()
+            this.typeHelper.isSpatial() ||
+            this.typeHelper.isBigInt()
         );
     }
 
@@ -140,7 +141,10 @@ export class AttributeAdapter {
     }
 
     isEventPayloadField(): boolean {
-        return this.typeHelper.isEnum() || this.typeHelper.isSpatial() || this.typeHelper.isScalar();
+        return (
+            (this.typeHelper.isEnum() || this.typeHelper.isSpatial() || this.typeHelper.isScalar()) &&
+            !this.isCustomResolvable()
+        );
     }
 
     isSubscriptionConnectedRelationshipField(): boolean {
@@ -158,7 +162,7 @@ export class AttributeAdapter {
             return false;
         }
 
-        if (this.timestampCreateIsGenerated()) {
+        if (this.timestampCreateIsGenerated() || this.populatedByCreateIsGenerated()) {
             return false;
         }
 
@@ -200,7 +204,8 @@ export class AttributeAdapter {
         return (
             this.isNonGeneratedField() &&
             this.annotations.settable?.onCreate !== false &&
-            !this.timestampCreateIsGenerated()
+            !this.timestampCreateIsGenerated() &&
+            !this.populatedByCreateIsGenerated()
         );
     }
 
@@ -208,7 +213,8 @@ export class AttributeAdapter {
         return (
             this.isNonGeneratedField() &&
             this.annotations.settable?.onUpdate !== false &&
-            !this.timestampUpdateIsGenerated()
+            !this.timestampUpdateIsGenerated() &&
+            !this.populatedByUpdateIsGenerated()
         );
     }
 
@@ -227,12 +233,27 @@ export class AttributeAdapter {
         return false;
     }
 
+    populatedByCreateIsGenerated(): boolean {
+        if (!this.annotations.populatedBy) {
+            // The populatedBy directive is not set on the field
+            return false;
+        }
+
+        if (this.annotations.populatedBy.operations.includes("CREATE")) {
+            // The populatedBy directive is set to generate on create
+            return true;
+        }
+
+        // The populatedBy directive is not set to generate on create
+        return false;
+    }
+
     isNonGeneratedField(): boolean {
         if (this.isCypher() || this.isCustomResolvable()) {
             return false;
         }
 
-        if (this.annotations.id || this.annotations.populatedBy) {
+        if (this.annotations.id) {
             return false;
         }
 
@@ -255,6 +276,21 @@ export class AttributeAdapter {
         }
 
         // The timestamp directive is not set to generate on update
+        return false;
+    }
+
+    populatedByUpdateIsGenerated(): boolean {
+        if (!this.annotations.populatedBy) {
+            // The populatedBy directive is not set on the field
+            return false;
+        }
+
+        if (this.annotations.populatedBy.operations.includes("UPDATE")) {
+            // The populatedBy directive is set to generate on update
+            return true;
+        }
+
+        // The populatedBy directive is not set to generate on update
         return false;
     }