+++ packages/graphql/src/translate/translate-update.ts
@@ -32,7 +32,7 @@ import { createConnectOrCreateAndParams } from "./create-connect-or-create-and-p
 import createCreateAndParams from "./create-create-and-params";
 import createDeleteAndParams from "./create-delete-and-params";
 import createDisconnectAndParams from "./create-disconnect-and-params";
-import createRelationshipValidationStr from "./create-relationship-validation-string";
+import { createRelationshipValidationString } from "./create-relationship-validation-string";
 import createSetRelationshipPropertiesAndParams from "./create-set-relationship-properties-and-params";
 import createUpdateAndParams from "./create-update-and-params";
 import { QueryASTContext, QueryASTEnv } from "./queryAST/ast/QueryASTContext";
@@ -71,9 +71,17 @@ export default async function translateUpdate({
     const createStrs: string[] = [];
     let deleteStr = "";
     const assumeReconnecting = Boolean(connectInput) && Boolean(disconnectInput);
-    const matchNode = new Cypher.NamedNode(varName, { labels: node.getLabels(context) });
+    const matchNode = new Cypher.NamedNode(varName);
     const where = resolveTree.args.where as GraphQLWhereArg | undefined;
-    const topLevelMatch = translateTopLevelMatch({ matchNode, node, context, operation: "UPDATE", where });
+    const matchPattern = new Cypher.Pattern(matchNode, { labels: node.getLabels(context) });
+    const topLevelMatch = translateTopLevelMatch({
+        matchNode,
+        matchPattern,
+        node,
+        context,
+        operation: "UPDATE",
+        where,
+    });
     matchAndWhereStr = topLevelMatch.cypher;
     let cypherParams = topLevelMatch.params;
 
@@ -462,7 +470,7 @@ export default async function translateUpdate({
         ? Cypher.concat(...queryASTResult.clauses)
         : new Cypher.Return(new Cypher.Literal("Query cannot conclude with CALL"));
 
-    const relationshipValidationStr = createRelationshipValidationStr({ node, context, varName });
+    const relationshipValidationStr = createRelationshipValidationString({ node, context, varName });
 
     const updateQuery = new Cypher.Raw((env: Cypher.Environment) => {
         const cypher = [