+++ packages/graphql/src/types/index.ts
@@ -27,6 +27,7 @@ import type { Integer } from "neo4j-driver";
 import type { RelationshipNestedOperationsOption, RelationshipQueryDirectionOption } from "../constants";
 import type { Neo4jGraphQLSchemaModel } from "../schema-model/Neo4jGraphQLSchemaModel";
 import type { DefaultAnnotationValue } from "../schema-model/annotation/DefaultAnnotation";
+import type { VectorField } from "../schema-model/annotation/VectorAnnotation";
 import type { JwtPayload } from "./jwt-payload";
 import type { Neo4jGraphQLContext } from "./neo4j-graphql-context";
 
@@ -47,6 +48,14 @@ export type FulltextContext = {
     queryName: string | undefined;
     indexName: string | undefined; // TODO: not undefined once name is removed.
     scoreVariable: Cypher.Variable;
+};
+
+export type VectorContext = {
+    index: VectorField;
+    queryName: string;
+    queryType: string;
+    scoreVariable: Cypher.Variable;
+    vectorSettings: Neo4jVectorSettings;
 };
 
 export type FullText = {
@@ -256,8 +265,11 @@ export type CallbackOperations = "CREATE" | "UPDATE";
   Object keys and enum values map to values at https://neo4j.com/docs/cypher-manual/current/query-tuning/query-options/#cypher-query-options
 */
 export interface CypherQueryOptions {
-    runtime?: "interpreted" | "slotted" | "pipelined";
+    runtime?: "interpreted" | "slotted" | "pipelined" | "parallel";
     planner?: "cost" | "idp" | "dp";
+    /**
+     * @deprecated The Cypher query option `connectComponentsPlanner` is deprecated and will be removed without a replacement. https://neo4j.com/docs/cypher-manual/current/planning-and-tuning/query-tuning/#cypher-connect-components-planner
+     */
     connectComponentsPlanner?: "greedy" | "idp";
     updateStrategy?: "default" | "eager";
     expressionEngine?: "default" | "interpreted" | "compiled";
@@ -388,7 +400,7 @@ export interface Neo4jGraphQLSubscriptionsEngine {
     close(): void;
 }
 
-export type CallbackReturnValue = string | number | boolean | undefined | null;
+export type CallbackReturnValue = string | number | boolean | undefined | null | Array<CallbackReturnValue>;
 
 export type Neo4jGraphQLCallback = (
     parent: Record<string, unknown>,
@@ -423,6 +435,32 @@ export interface Neo4jAuthorizationSettings {
     verify?: boolean;
     verifyOptions?: JWTVerifyOptions;
 }
+
+export interface Neo4jVectorSettings {
+    ["VertexAI"]?: {
+        token: string; // API access token.
+        projectId: string; // GCP project ID.
+        model?: string; // The name of the model you want to invoke.
+        region?: string; // GCP region where to send the API requests.
+    };
+    ["OpenAI"]?: {
+        token: string; // API access token.
+        model?: string; // The name of the model you want to invoke.
+        dimensions?: number; // The number of dimensions you want to reduce the vector to. Only supported for certain models.
+    };
+    ["AzureOpenAI"]?: {
+        token: string; // API access token.
+        resource: string; // The name of the resource to which the model has been deployed.
+        deployment: string; // The name of the model deployment.
+    };
+    ["Bedrock"]?: {
+        accessKeyId: string; // AWS access key ID.
+        secretAccessKey: string; // AWS secret key.
+        model?: string; // The name of the model you want to invoke.
+        region?: string; // AWS region where to send the API requests.
+    };
+}
+
 export interface RemoteJWKS {
     url: string | URL;
     options?: RemoteJWKSetOptions;
@@ -435,6 +473,18 @@ export type Neo4jFeaturesSettings = {
     populatedBy?: Neo4jPopulatedBySettings;
     authorization?: Neo4jAuthorizationSettings;
     subscriptions?: Neo4jGraphQLSubscriptionsEngine | boolean;
+    /** If set to `true`, removes `@neo4j/graphql` fields that are marked as deprecated to reduce schema size.
+     *
+     * NOTE: this will not remove user defined deprecated fields
+     **/
+    excludeDeprecatedFields?: {
+        bookmark?: boolean;
+        negationFilters?: boolean;
+        arrayFilters?: boolean;
+        stringAggregation?: boolean;
+        aggregationFilters?: boolean;
+    };
+    vector?: Neo4jVectorSettings;
 };
 
 /** Parsed features used in context */